Move Lsa handles to Interop code
