+++ src/Common/src/Interop/Windows/Advapi32/Interop.LsaLookupSids.cs
@@ -17,6 +17,6 @@ internal static partial class Interop
             IntPtr[] sids,
             out SafeLsaMemoryHandle referencedDomains,
             out SafeLsaMemoryHandle names
-            );
+        );
     }
 }
