+++ src/Common/src/Interop/Windows/SspiCli/Interop.LSA_STRING.cs
