+++ src/Common/src/Interop/Windows/SspiCli/Interop.LSA_UNICODE_STRING.cs
