+++ src/Common/src/Microsoft/Win32/SafeHandles/SafeTokenHandle.cs
