+++ src/System.Security.Principal.Windows/src/System.Security.Principal.Windows.csproj
@@ -10,7 +10,6 @@
   </PropertyGroup>
   <ItemGroup Condition="($(TargetGroup.StartsWith('netcoreapp')) or '$(TargetGroup)' == 'uap') AND '$(TargetsWindows)' == 'true'">
     <Compile Include="Microsoft\Win32\SafeHandles\SafeAccessTokenHandle.cs" />
-    <Compile Include="Microsoft\Win32\SafeHandles\SafeSecurityHandles.cs" />
     <Compile Include="System\Security\Principal\IdentityNotMappedException.cs" />
     <Compile Include="System\Security\Principal\IdentityReference.cs" />
     <Compile Include="System\Security\Principal\IRCollection.cs" />
@@ -146,11 +145,11 @@
     <Compile Include="$(CommonPath)\Interop\Windows\SspiCli\Interop.LsaLookupAuthenticationPackage.cs">
       <Link>Common\Interop\Windows\SspiCli\Interop.LsaLookupAuthenticationPackage.cs</Link>
     </Compile>
-    <Compile Include="$(CommonPath)\Interop\Windows\SspiCli\Interop.LsaString.cs">
-      <Link>Common\Interop\Windows\SspiCli\Interop.LsaString.cs</Link>
+    <Compile Include="$(CommonPath)\Interop\Windows\SspiCli\Interop.LSA_STRING.cs">
+      <Link>Common\Interop\Windows\SspiCli\Interop.LSA_STRING.cs</Link>
     </Compile>
-    <Compile Include="$(CommonPath)\Interop\Windows\SspiCli\Interop.LsaUnicodeString.cs">
-      <Link>Common\Interop\Windows\SspiCli\Interop.LsaUnicodeString.cs</Link>
+    <Compile Include="$(CommonPath)\Interop\Windows\SspiCli\Interop.LSA_UNICODE_STRING.cs">
+      <Link>Common\Interop\Windows\SspiCli\Interop.LSA_UNICODE_STRING.cs</Link>
     </Compile>
     <Compile Include="$(CommonPath)\Interop\Windows\SspiCli\Interop.QuotaLimits.cs">
       <Link>Common\Interop\Windows\SspiCli\Interop.QuotaLimits.cs</Link>
@@ -169,6 +168,15 @@
     </Compile>
     <Compile Include="$(CommonPath)\Microsoft\Win32\SafeHandles\SafeLsaHandle.cs">
       <Link>Common\Microsoft\Win32\SafeHandles\SafeLsaHandle.cs</Link>
+    </Compile>
+    <Compile Include="$(CommonPath)\Microsoft\Win32\SafeHandles\SafeLsaMemoryHandle.cs">
+      <Link>Common\Microsoft\Win32\SafeHandles\SafeLsaMemoryHandle.cs</Link>
+    </Compile>
+    <Compile Include="$(CommonPath)\Microsoft\Win32\SafeHandles\SafeLsaPolicyHandle.cs">
+      <Link>Common\Microsoft\Win32\SafeHandles\SafeLsaPolicyHandle.cs</Link>
+    </Compile>
+    <Compile Include="$(CommonPath)\Microsoft\Win32\SafeHandles\SafeLsaReturnBufferHandle.cs">
+      <Link>Common\Microsoft\Win32\SafeHandles\SafeLsaReturnBufferHandle.cs</Link>
     </Compile>
   </ItemGroup>
   <ItemGroup Condition="$(TargetGroup.StartsWith('netcoreapp')) AND '$(TargetsWindows)' == 'true'">
